Optimizations and Additional Functionalities. Added utilities app. Created custom_tags.py in utilities app. Moved decorators.py in utilities app. Allows addition of classes for svgs (company_logo1, company_logo2, company_logo3). Separate the rendering of account type form sfrom the account_type_selection.html. Added render_account_type_radio template tags in custom_tags.py. Added Flowbite CDN. Added DarkMode in progress.
