+++ authentication_feature/templates/account_type_selection.html
@@ -1,22 +1,18 @@
-{% extends "base1.html" %} {% block title %}Select Account Type{% endblock %}
+{% extends "base1.html" %} {% load custom_tags %} {% block title %}Select
+Account Type{% endblock %}
 <!-- prettier-ignore -->
 {% block content %}
-<div class="container mx-auto mt-24">
-	<span class="flex justify-center">
-		{% include 'svgs/company_logo1.html' %}
-	</span>
-	<form
-		class="border rounded bg-huntify-gray-3 text-center"
-		method="POST">
-		<span class="font-inter text-2xl">
-			Books and view property listings that you like
-		</span>
-	</form>
+<div class="flex justify-center mt-24">
+	<div class="text-center">
+		<!-- Logo -->
+		<!-- prettier-ignore -->
+		{% include 'svgs/company_logo1.html' with class="w-[198px] h-[192px] mx-auto" %}
+		<!-- Header -->
+		<h2 class="text-4xl font-bold mt-12 font-poppins text-huntify-gray-1">
+			Select user type
+		</h2>
+		<!-- User Type Selection -->
+		{% render_account_type_form form %}
+	</div>
 </div>
-{% comment %}
-<h1>Select Account Type</h1>
-<form method="post">
-	{% csrf_token %} {{ form.as_p }}
-	<button type="submit">Next</button>
-</form>
-{% endcomment %} {% endblock %}
+{% endblock %}
