+++ templates/base1.html
@@ -3,5 +3,7 @@
 	{% include 'partials/head.html' %}
 	<body>
 		{% block content %} {% endblock %}
+		<!-- Flowbite JS CDN -->
+		<script src="https://cdn.jsdelivr.net/npm/flowbite@2.5.2/dist/flowbite.min.js"></script>
 	</body>
 </html>
