+++ templates/partials/head.html
@@ -6,6 +6,10 @@
 		content="width=device-width, initial-scale=1.0" />
 	<!-- Tailwind CSS CDN -->
 	<script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio,line-clamp,container-queries"></script>
+	<!-- Flowbite CSS CDN -->
+	<link
+		href="https://cdn.jsdelivr.net/npm/flowbite@2.5.2/dist/flowbite.min.css"
+		rel="stylesheet" />
 	<!-- Google Fonts -->
 	<link
 		href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=Poppins:wght@400;500;600;700&display=swap"
@@ -34,6 +38,7 @@
 					},
 				},
 			},
+			darkMode: "class",
 		}
 	</script>
 	<title>{% block title %}Huntify{% endblock %}</title>
